Move whole app to main page.tsx as we only have one page
